Setup eslint and prettier
Eslint can be run with `yarn lint:all`, you can autofix some errors with
`yarn lint:fix`.

This also adds some configuration for prettier, a plugin that will
automatically format your code to conform to the styleguide when saving
a file. This is configured as a shared VS Code setting.
if using a different editor, this will likely have to be a custom config.