Implement add and remove favourites actions and reducer handlers
Now the `Pointer` component will dispatch actions to add or remove the
favourite state, instead of managing this via its internal `this.state`
property. Now that redux is handling this, the `FavouritesList`
component is automatically updated since it is connected to the redux
store and is now aware of the updates.