Remove function for Favourites List
Connect the favourites list with the `removeFavourite` action and bind
this to a remove button.

This works for the default data, but quickly fails, due to the data
structure being an Array and we use the `index` to update the state.

Next we will change the data structure in order to make the state easier
to work with.