Configure Jest to handle component snapshot testing
Here we use `enzyme`, a library that can mount React components and test
them in isolation.

We need `jest-css-modules` to prevent a parsing error, since we are using
`postcss`.

Also included is the `enzyme-to-json` serializer which gives us *much*
more readable snapshots.

`enzyme-adapter-react-16` is used to handle React v16.

Finally a `spec-helper` file is added, which is automatically run before
the test suite. Here is information to polyfill `requestAnimationFrame`
and configure enzyme to use the React v16 adapter.

`yarn add --dev enzyme enzyme-adapter-react-16 enzyme-to-json jest-css-modules`