Test the dumb `Map` component.
Checks the snapshots are built for a map with and without points data.

In order to do this, we must also export the Map that is not connected.
(We use `*` to export anything else that might be there too).

Enzyme does not know about redux yet, so we want to test the
non-connected version of the Map for now.

Next we will test the connected version by created a mock redux store.